refactor(auth): rename authentication to auth
